+++ src/testRunner/unittests/tsserver/projects.ts
@@ -1075,121 +1075,6 @@ namespace ts.projectSystem {
             assert.equal(options.outDir, "C:/a/b", "");
         });
 
-        it("dynamic file without external project", () => {
-            const file: File = {
-                path: "^walkThroughSnippet:/Users/UserName/projects/someProject/out/someFile#1.js",
-                content: "var x = 10;"
-            };
-            const host = createServerHost([libFile], { useCaseSensitiveFileNames: true });
-            const projectService = createProjectService(host);
-            projectService.setCompilerOptionsForInferredProjects({
-                module: ModuleKind.CommonJS,
-                allowJs: true,
-                allowSyntheticDefaultImports: true,
-                allowNonTsExtensions: true
-            });
-            projectService.openClientFile(file.path, "var x = 10;");
-
-            projectService.checkNumberOfProjects({ inferredProjects: 1 });
-            const project = projectService.inferredProjects[0];
-            checkProjectRootFiles(project, [file.path]);
-            checkProjectActualFiles(project, [file.path, libFile.path]);
-            verifyDynamic(projectService, `/${file.path}`);
-
-            assert.strictEqual(projectService.ensureDefaultProjectForFile(server.toNormalizedPath(file.path)), project);
-            const indexOfX = file.content.indexOf("x");
-            assert.deepEqual(project.getLanguageService(/*ensureSynchronized*/ true).getQuickInfoAtPosition(file.path, indexOfX), {
-                kind: ScriptElementKind.variableElement,
-                kindModifiers: "",
-                textSpan: { start: indexOfX, length: 1 },
-                displayParts: [
-                    { text: "var", kind: "keyword" },
-                    { text: " ", kind: "space" },
-                    { text: "x", kind: "localName" },
-                    { text: ":", kind: "punctuation" },
-                    { text: " ", kind: "space" },
-                    { text: "number", kind: "keyword" }
-                ],
-                documentation: [],
-                tags: undefined,
-            });
-        });
-
-        it("dynamic file with reference paths without external project", () => {
-            const file: File = {
-                path: "^walkThroughSnippet:/Users/UserName/projects/someProject/out/someFile#1.js",
-                content: `/// <reference path="../../../../../../typings/@epic/Core.d.ts" />
-/// <reference path="../../../../../../typings/@epic/Shell.d.ts" />
-var x = 10;`
-            };
-            const host = createServerHost([libFile]);
-            const projectService = createProjectService(host);
-            projectService.openClientFile(file.path, file.content);
-            verifyDynamic(projectService, projectService.toPath(file.path));
-
-            projectService.checkNumberOfProjects({ inferredProjects: 1 });
-            const project = projectService.inferredProjects[0];
-            checkProjectRootFiles(project, [file.path]);
-            checkProjectActualFiles(project, [file.path, libFile.path]);
-        });
-
-        describe("dynamic file with projectRootPath", () => {
-            const file: File = {
-                path: "^walkThroughSnippet:/Users/UserName/projects/someProject/out/someFile#1.js",
-                content: "var x = 10;"
-            };
-            const configFile: File = {
-                path: `${tscWatch.projectRoot}/tsconfig.json`,
-                content: "{}"
-            };
-            const configProjectFile: File = {
-                path: `${tscWatch.projectRoot}/a.ts`,
-                content: "let y = 10;"
-            };
-            it("with useInferredProjectPerProjectRoot", () => {
-                const host = createServerHost([libFile, configFile, configProjectFile], { useCaseSensitiveFileNames: true });
-                const session = createSession(host, { useInferredProjectPerProjectRoot: true });
-                openFilesForSession([{ file: file.path, projectRootPath: tscWatch.projectRoot }], session);
-
-                const projectService = session.getProjectService();
-                checkNumberOfProjects(projectService, { inferredProjects: 1 });
-                checkProjectActualFiles(projectService.inferredProjects[0], [file.path, libFile.path]);
-                verifyDynamic(projectService, `${tscWatch.projectRoot}/${file.path}`);
-
-                session.executeCommandSeq<protocol.OutliningSpansRequest>({
-                    command: protocol.CommandTypes.GetOutliningSpans,
-                    arguments: {
-                        file: file.path
-                    }
-                });
-
-                // Without project root
-                const file2Path = file.path.replace("#1", "#2");
-                projectService.openClientFile(file2Path, file.content);
-                checkNumberOfProjects(projectService, { inferredProjects: 2 });
-                checkProjectActualFiles(projectService.inferredProjects[0], [file.path, libFile.path]);
-                checkProjectActualFiles(projectService.inferredProjects[1], [file2Path, libFile.path]);
-            });
-
-            it("fails when useInferredProjectPerProjectRoot is false", () => {
-                const host = createServerHost([libFile, configFile, configProjectFile], { useCaseSensitiveFileNames: true });
-                const projectService = createProjectService(host);
-                try {
-                    projectService.openClientFile(file.path, file.content, /*scriptKind*/ undefined, tscWatch.projectRoot);
-                }
-                catch (e) {
-                    assert.strictEqual(
-                        e.message.replace(/\r?\n/, "\n"),
-                        `Debug Failure. False expression: \nVerbose Debug Information: {"fileName":"^walkThroughSnippet:/Users/UserName/projects/someProject/out/someFile#1.js","currentDirectory":"/user/username/projects/myproject","hostCurrentDirectory":"/","openKeys":[]}\nDynamic files must always be opened with service's current directory or service should support inferred project per projectRootPath.`
-                    );
-                }
-                const file2Path = file.path.replace("#1", "#2");
-                projectService.openClientFile(file2Path, file.content);
-                projectService.checkNumberOfProjects({ inferredProjects: 1 });
-                checkProjectActualFiles(projectService.inferredProjects[0], [file2Path, libFile.path]);
-            });
-        });
-
         it("files opened, closed affecting multiple projects", () => {
             const file: File = {
                 path: "/a/b/projects/config/file.ts",
